Use rust ufcs method call in extern trait specs
