--- prusti-tests/tests/verify/pass/extern-spec/traits/generics-3.rs
@@ -1,35 +1,0 @@
-// ignore-test
-// TODO: 
-extern crate prusti_contracts;
-use prusti_contracts::*;
-
-trait MyTrait<T> {
-    fn get_value(&self) -> T;
-}
-
-#[extern_spec]
-trait MyTrait<#[concrete] i32 > {
-    #[ensures(result == 42)]
-    fn get_value(&self) -> i32;
-}
-
-#[extern_spec]
-trait MyTrait<#[concrete] u32 > {
-    #[ensures(result == 43)]
-    fn get_value(&self) -> u32;
-}
-
-struct Impl;
-impl MyTrait<i32> for Impl {
-    fn get_value(&self) -> i32 { 42 }
-}
-impl MyTrait<u32> for Impl {
-    fn get_value(&self) -> u32 { 43 }
-}
-
-
-fn main() {
-    let s = Impl{};
-    assert!(MyTrait::<i32>::get_value(&s) == 42);
-    assert!(MyTrait::<u32>::get_value(&s) == 43);
-}
